Refactoring code and adding rating emojis
