+++ src/discord/commands/stalk.ts
@@ -3,8 +3,11 @@ import type { Command } from "../type";
 import { db } from "$db/index";
 import { CFApiFactory } from "$src/codeforces/client";
 import CliTable3 from "cli-table3";
+import { wrapText, formatRating } from "$src/lib/utils";
+import { getNavButtons } from "$src/lib/discordUtils";
 
 const INT_MAX = 2147483647;
+const collectorTime = 60000 * 3;
 
 export const stalkCmd: Command = {
     info: new SlashCommandBuilder()
@@ -22,8 +25,9 @@ export const stalkCmd: Command = {
         ),
 
     async execute(msg) {
-        await msg.deferReply(); // defer reply to avoid timeout
+        await msg.deferReply(); // Defer to await
 
+        // Get user ID if mentioned
         const mention = msg.options.getUser('user');
         const selectedUser = mention ? mention : msg.user;
 
@@ -31,6 +35,7 @@ export const stalkCmd: Command = {
         if (!user)
             return msg.editReply("User has not registered their Codeforces handle!");
 
+        // Fetch user submissions
         const cfApi = CFApiFactory.get();
         const allUserSubmissions = await cfApi.getUserSubmissions(user.handle, 10000);
 
@@ -38,6 +43,7 @@ export const stalkCmd: Command = {
             return msg.editReply("Participate in some contests!");
         }
 
+        // Sort submissions
         const sortByRating = msg.options.getBoolean("sort");
         if (sortByRating) {
             allUserSubmissions.sort((a, b) => {
@@ -57,6 +63,7 @@ export const stalkCmd: Command = {
             });
         }
 
+        // Filter Unique and Contest only
         const onlyContest = msg.options.getBoolean("contest") != null ? msg.options.getBoolean("contest") : false;
         console.log(onlyContest);
         const processedProblems = new Set<string>();
@@ -74,67 +81,33 @@ export const stalkCmd: Command = {
         const chunkSize = 10;
         const totalPages = Math.ceil(filteredSubmissions.length / chunkSize);
 
-        const wrapText = (text: string, maxWidth: number): string => {
-            const words = text.split(' ');
-            let line = '';
-            let result = '';
-            for (let word of words) {
-                if (line.length + word.length > maxWidth) {
-                    result += line + '\n';  // wrap the text
-                    line = word + ' ';
-                } else {
-                    line += word + ' ';
-                }
-            }
-            result += line.trim(); // Add the last line
-            return result;
-        };
-
-        const getEmojiForRating = (rating: number): string => {
-            if (rating < 1200) return "🔘";
-            if (rating < 1400) return "🟢";
-            if (rating < 1600) return "🩵";
-            if (rating < 1900) return "🔵";
-            if (rating < 2100) return "🟣";
-            if (rating < 2300) return "🟡";
-            if (rating < 2400) return "🟠";
-            if (rating < 2600) return "🔴";
-            if (rating < 3000) return "🔥";
-            if (rating < 3500) return "✨";
-            return "👑";
-        };
-
-        const formatRating = (rating: number): string => {
-            const emoji = getEmojiForRating(rating);
-            const ratingString = rating.toString();
-
-            return `${emoji} ${ratingString}`;
-        };
-
         const createSubmissionTable = (page: number): string => {
             const start = page * chunkSize;
             const end = Math.min(start + chunkSize, filteredSubmissions.length);
+
             const nameWidth = 19;
             const ratingWidth = 11;
             const submissionTimeWidth = 19;
             const table = new CliTable3({
                 head: ['Problem Name', 'Rating', 'Submission Time'],
                 style: {
-                    head: [], // disable colors in header cells
-                    border: [], // disable colors for the border
+                    head: [],
+                    border: [],
                 },
                 colAligns: ['center', 'center', 'center'],
-                colWidths: [nameWidth, ratingWidth, submissionTimeWidth], // set the widths of each column (optional)
+                colWidths: [nameWidth, ratingWidth, submissionTimeWidth],
             });
 
             for (let i = start; i < end; i++) {
                 const problem = filteredSubmissions[i].problem;
-                const rating = problem.rating ? formatRating(problem.rating) : '?';
-                const name = wrapText(problem.name, nameWidth); // Wrap the text if it's too long
+
+                const rating = problem.rating ? formatRating(problem.rating, ratingWidth) : '?';
+                
+                const name = wrapText(problem.name, nameWidth);
 
                 const timeAgo = (timeInS: number): string => {
                     const now = Date.now();
-                    const diff = now - timeInS * 1000; // Convert to milliseconds
+                    const diff = now - timeInS * 1000;
                     const seconds = Math.floor(diff / 1000);
                     const minutes = Math.floor(seconds / 60);
                     const hours = Math.floor(minutes / 60);
@@ -150,8 +123,8 @@ export const stalkCmd: Command = {
                         return `${seconds} second(s) ago`;
                     }
                 };
-
                 const submissionTime = wrapText(timeAgo(filteredSubmissions[i].creationTimeSeconds), submissionTimeWidth);
+
                 table.push([name, rating.toString(), submissionTime]);
             }
 
@@ -161,19 +134,7 @@ export const stalkCmd: Command = {
         let currentPage = 0;
         const tableMsg = createSubmissionTable(currentPage);
 
-        const row = new ActionRowBuilder<ButtonBuilder>()
-            .addComponents(
-                new ButtonBuilder()
-                    .setCustomId('prev')
-                    .setLabel('Previous')
-                    .setStyle(ButtonStyle.Primary)
-                    .setDisabled(currentPage === 0),
-                new ButtonBuilder()
-                    .setCustomId('next')
-                    .setLabel('Next')
-                    .setStyle(ButtonStyle.Primary)
-                    .setDisabled(currentPage === totalPages - 1),
-            );
+        const row = getNavButtons(currentPage, totalPages);
 
         await msg.editReply({
             content: `\`\`\`User ${user.handle} Submissions\n\n${tableMsg}\`\`\``,
@@ -181,7 +142,7 @@ export const stalkCmd: Command = {
         });
 
         const filter = (i: any) => i.user.id === msg.user.id;
-        const collector = msg.channel.createMessageComponentCollector({ filter, time: 60000 });
+        const collector = msg.channel.createMessageComponentCollector({ filter, time: collectorTime });
 
         collector.on('collect', async (interaction) => {
             if (interaction.customId === 'prev') {
@@ -192,19 +153,7 @@ export const stalkCmd: Command = {
 
             const updatedTableMsg = createSubmissionTable(currentPage);
 
-            const updatedRow = new ActionRowBuilder<ButtonBuilder>()
-                .addComponents(
-                    new ButtonBuilder()
-                        .setCustomId('prev')
-                        .setLabel('Previous')
-                        .setStyle(ButtonStyle.Primary)
-                        .setDisabled(currentPage === 0),
-                    new ButtonBuilder()
-                        .setCustomId('next')
-                        .setLabel('Next')
-                        .setStyle(ButtonStyle.Primary)
-                        .setDisabled(currentPage === totalPages - 1),
-                );
+            const updatedRow = getNavButtons(currentPage, totalPages);
 
             await interaction.update({
                 content: `\`\`\`User ${user.handle} Submissions\n\n${updatedTableMsg}\`\`\``,
