+++ src/lib/discordUtils.ts
@@ -1,0 +1,18 @@
+import { ButtonBuilder, ActionRowBuilder, ButtonStyle } from "discord.js";
+
+export const getNavButtons = (currentPage: number, totalPages: number) => {
+    const row = new ActionRowBuilder<ButtonBuilder>()
+        .addComponents(
+            new ButtonBuilder()
+                .setCustomId('prev')
+                .setLabel('Previous')
+                .setStyle(ButtonStyle.Primary)
+                .setDisabled(currentPage === 0),
+            new ButtonBuilder()
+                .setCustomId('next')
+                .setLabel('Next')
+                .setStyle(ButtonStyle.Primary)
+                .setDisabled(currentPage === totalPages - 1),
+        );
+    return row;
+};
